llm/pyproject.toml
[project]
name = "llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "langsmith>=0.4.15",
    "python-dotenv>=1.1.1",
]
